requirements.txt
alembic==1.8.0
asyncpg==0.26.0
fastapi==0.78.0
pydantic[dotenv]
SQLAlchemy==1.4.37
uvicorn
gunicorn
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart